MockData moved to examples
