+++ examples/MockData/Data.ts
@@ -20,7 +20,8 @@ export const data = [
 {argument: new Date('06/19/2016'), value: 50},
 {argument: new Date('06/20/2016'), value: 29},
 {argument: new Date('06/21/2016'), value: 36},
-{argument: new Date('06/22/2016'), value: 32}];
+{argument: new Date('06/22/2016'), value: 32}
+];
 
 export const data2 = [
 {argument: new Date('06/01/2016'), value: 25},
@@ -44,7 +45,8 @@ export const data2 = [
 {argument: new Date('06/19/2016'), value: 32},
 {argument: new Date('06/20/2016'), value: 25},
 {argument: new Date('06/21/2016'), value: 47},
-{argument: new Date('06/22/2016'), value: 53}];
+{argument: new Date('06/22/2016'), value: 53}
+];
 
 export const updatedData = [
 {argument: new Date('06/10/2016'), value: 25},
@@ -67,7 +69,8 @@ export const updatedData = [
 {argument: new Date('06/27/2016'), value: 39},
 {argument: new Date('06/28/2016'), value: 32},
 {argument: new Date('06/29/2016'), value: 25},
-{argument: new Date('06/30/2016'), value: 47},];
+{argument: new Date('06/30/2016'), value: 47}
+];
 
 export const updatedData1 = [
 {argument: new Date('07/10/2016'), value: 25},
@@ -90,7 +93,8 @@ export const updatedData1 = [
 {argument: new Date('07/27/2016'), value: 39},
 {argument: new Date('07/28/2016'), value: 32},
 {argument: new Date('07/29/2016'), value: 25},
-{argument: new Date('07/30/2016'), value: 47},];
+{argument: new Date('07/30/2016'), value: 47}
+];
 
 export const sData1 = {
     name: 'Data 1',
