+++ examples/MockData/barChart.ts
